requirements.txt
Django==2.0
Markdown==3.0.1
Pillow==5.3.0
pydocs==0.2
pytz==2018.7
redis==3.0.1
celery==3.1.25
django-celery==3.2.2
django-filter==2.0.0
djangorestframework==3.9.0
django-extensions==2.1.4
django-debug-toolbar==1.11
docutils==0.14
gunicorn==19.9.0
#uWSGI==2.0.10
fcm-django==0.2.20
psycopg2==2.7.6.1
